+++ src/api/mod.rs
@@ -5,32 +5,40 @@ pub mod control;
 
 use std::fmt::Debug;
 
-use actix::Addr;
 use futures::Future;
+use medea_client_api_proto::Command;
 
-use crate::{
-    api::client::rpc_connection::{
-        CommandMessage, RpcConnectionClosed, RpcConnectionEstablished,
-    },
-    log::prelude::*,
-    signalling::Room,
+use crate::api::client::rpc_connection::{
+    RpcConnectionClosed, RpcConnectionEstablished,
 };
 
 #[cfg_attr(test, mockall::automock)]
 pub trait RpcServer: Debug + Send {
+    /// Send signal of new [`RpcConnection`] being established with specified
+    /// [`Member`]. Transport should consider dropping connection if message
+    /// result is err.
+    ///
+    /// [`Member`]: crate::signalling::elements::member::Member
     fn send_established(
         &self,
         msg: RpcConnectionEstablished,
     ) -> Box<dyn Future<Item = (), Error = ()>>;
 
+    /// Send signal of existing [`RpcConnection`] of specified [`Member`] being
+    /// closed. Cannot fail.
+    ///
+    /// [`Member`]: crate::signalling::elements::member::Member
     fn send_closed(
         &self,
         msg: RpcConnectionClosed,
     ) -> Box<dyn Future<Item = (), Error = ()>>;
 
+    /// Sends [`Command`]. Cannot fail
+    ///
+    /// [`Command`]:
     fn send_command(
         &self,
-        msg: CommandMessage,
+        msg: Command,
     ) -> Box<dyn Future<Item = (), Error = ()>>;
 }
 
@@ -38,52 +46,5 @@ pub trait RpcServer: Debug + Send {
 impl Debug for MockRpcServer {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         write!(f, "MockRpcServer")
-    }
-}
-
-impl RpcServer for Addr<Room> {
-    fn send_established(
-        &self,
-        msg: RpcConnectionEstablished,
-    ) -> Box<dyn Future<Item = (), Error = ()>> {
-        Box::new(
-            self.send(msg)
-                .map_err(|err| {
-                    error!(
-                        "Failed to send RpcConnectionEstablished cause {:?}",
-                        err,
-                    );
-                })
-                .and_then(|result| result),
-        )
-    }
-
-    fn send_closed(
-        &self,
-        msg: RpcConnectionClosed,
-    ) -> Box<dyn Future<Item = (), Error = ()>> {
-        Box::new(
-            self.send(msg)
-                .map_err(|err| {
-                    error!(
-                        "Failed to send RpcConnectionClosed cause {:?}",
-                        err,
-                    );
-                })
-                .then(|_| Ok(())),
-        )
-    }
-
-    fn send_command(
-        &self,
-        msg: CommandMessage,
-    ) -> Box<dyn Future<Item = (), Error = ()>> {
-        Box::new(
-            self.send(msg)
-                .map_err(|err| {
-                    error!("Failed to send CommandMessage cause {:?}", err,);
-                })
-                .then(|_| Ok(())),
-        )
     }
 }
