+++ src/signalling/room.rs
@@ -6,12 +6,12 @@
 use std::collections::{HashMap, HashSet};
 
 use actix::{
-    fut::wrap_future, Actor, ActorFuture, AsyncContext, Context, Handler,
+    fut::wrap_future, Actor, ActorFuture, Addr, AsyncContext, Context, Handler,
     Message, ResponseActFuture, WrapFuture as _,
 };
 use derive_more::Display;
 use failure::Fail;
-use futures::future;
+use futures::future::{self, Future};
 use medea_client_api_proto::{
     Command, CommandHandler, Event, IceCandidate, PeerId, PeerMetrics, TrackId,
 };
@@ -35,6 +35,7 @@ use crate::{
             EndpointId, EndpointSpec, MemberId, MemberSpec, RoomId,
             TryFromElementError, WebRtcPlayId, WebRtcPublishId,
         },
+        RpcServer,
     },
     log::prelude::*,
     media::{
@@ -747,6 +748,58 @@ impl Room {
         self.members.insert_member(id, signalling_member);
 
         Ok(())
+    }
+}
+
+impl RpcServer for Addr<Room> {
+    /// Sends [`RpcConnectionEstablished`] message to [`Room`] actor propagating
+    /// errors.
+    fn send_established(
+        &self,
+        msg: RpcConnectionEstablished,
+    ) -> Box<dyn Future<Item = (), Error = ()>> {
+        Box::new(
+            self.send(msg)
+                .map_err(|err| {
+                    error!(
+                        "Failed to send RpcConnectionEstablished cause {:?}",
+                        err,
+                    );
+                })
+                .and_then(|result| result),
+        )
+    }
+
+    /// Sends [`RpcConnectionClosed`] message to [`Room`] actor ignoring any
+    /// errors.
+    fn send_closed(
+        &self,
+        msg: RpcConnectionClosed,
+    ) -> Box<dyn Future<Item = (), Error = ()>> {
+        Box::new(
+            self.send(msg)
+                .map_err(|err| {
+                    error!(
+                        "Failed to send RpcConnectionClosed cause {:?}",
+                        err,
+                    );
+                })
+                .then(|_| Ok(())),
+        )
+    }
+
+    /// Sends [`CommandMessage`] message to [`Room`] actor ignoring any errors.
+    fn send_command(
+        &self,
+        msg: Command,
+    ) -> Box<dyn Future<Item = (), Error = ()>> {
+        Box::new(
+            self.send(CommandMessage::from(msg))
+                .map_err(|err| {
+                    error!("Failed to send CommandMessage cause {:?}", err,);
+                })
+                .then(|_| Ok(())),
+        )
     }
 }
